rpi4/requirements.txt
pyserial==3.5
feetech-servo-sdk==1.0.0
numpy
pyzmq
opencv-python-headless
